refactored notifications menu list
