+++ client/src/components/NavBar/NotificationsMenu/NotificationsMenu.tsx
@@ -1,25 +1,12 @@
-import useStyles from './useStyles';
 import { useAuth } from '../../../context/useAuthContext';
-import {
-  Typography,
-  Button,
-  Badge,
-  MenuItem,
-  Menu,
-  Paper,
-  ListItemText,
-  ListItemAvatar,
-  Avatar,
-  Box,
-} from '@material-ui/core';
+import { Typography, Button, Badge, Menu, Box } from '@material-ui/core';
 import { Notification } from '../../../interface/Notification';
 import React, { useEffect, useState } from 'react';
 import { getAllUnreadNotifications, markAsRead } from '../../../helpers/APICalls/Notification';
+import NotificationsMenuItems from './NotificationsMenuItems/NotificationsMenuItems';
 
 export default function NotificationsMenu(): JSX.Element {
-  const classes = useStyles();
   const { loggedInUser } = useAuth();
-  const defaultLogoURL = 'https://res.cloudinary.com/dalisapxa/image/upload/v1630880229/DEV/logo_nywmrf.png';
 
   const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
 
@@ -83,43 +70,7 @@ export default function NotificationsMenu(): JSX.Element {
           />
 
           <Box borderColor={'#000000'} borderTop={5}>
-            <Paper>
-              {notifications &&
-                notifications.map((notification, index) => (
-                  <Box key={index} paddingLeft={'4px'}>
-                    <MenuItem onClick={handleClose}>
-                      <ListItemAvatar>
-                        <Avatar
-                          src={notification.context ? notification.context.profilePhotoURL : defaultLogoURL}
-                          variant="square"
-                          className={classes.notificationAvatar}
-                        />
-                      </ListItemAvatar>
-
-                      <ListItemText
-                        secondary={
-                          <Box>
-                            <Typography>
-                              <Box fontSize={12} fontWeight={600} color="#000000">
-                                {notification.title}
-                              </Box>
-                            </Typography>
-
-                            <Typography variant="button">
-                              <Box fontSize={10}>{notification.notificationType}</Box>
-                            </Typography>
-                            <Typography>
-                              <Box fontSize={12} fontWeight={600} color="#000000">
-                                {new Date(notification.createdAt).toLocaleDateString()}
-                              </Box>
-                            </Typography>
-                          </Box>
-                        }
-                      />
-                    </MenuItem>
-                  </Box>
-                ))}
-            </Paper>
+            <NotificationsMenuItems notifications={notifications} />
           </Box>
         </Menu>
       )}
